+++ src/components/SearchBar.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Input, Button, Dropdown } from '@stardust-ui/react';
-import '../SearchBar.css'
+import '../css/SearchBar.css'
+import { RadioIcons } from './RadioIcons';
 
 enum viewType{
   List= 'List',
@@ -53,6 +54,7 @@ export class SearchBar extends React.Component<ISearchBarProps, ISearchBarState>
   public render() {
     return(
       <div className="SearchBar">
+        <RadioIcons />
         <Dropdown inline items={inputItems} onSelectedChange={this.handleDropdownChange} placeholder="Select a view..." />
         <span id="search">
           <Input placeholder="Search..." icon="search" onChange={e => this.handleOnChange(e)}/>
